Incorporating reveiw comments - Moving mbed_rtos_types.h to public headers
