+++ rtos/include/rtos/EventFlags.h
@@ -26,7 +26,7 @@
 #include <cstddef>
 #include <stdint.h>
 #include "rtos/Kernel.h"
-#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/mbed_rtos_types.h"
 #include "rtos/internal/mbed_rtos1_types.h"
 #include "rtos/internal/mbed_rtos_storage.h"
