+++ rtos/include/rtos/Mail.h
@@ -28,7 +28,7 @@
 
 #include "rtos/Queue.h"
 #include "rtos/MemoryPool.h"
-#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/mbed_rtos_types.h"
 #include "rtos/internal/mbed_rtos_storage.h"
 #include "rtos/internal/mbed_rtos1_types.h"
