+++ rtos/include/rtos/MemoryPool.h
@@ -26,7 +26,7 @@
 #include <stdint.h>
 #include <string.h>
 
-#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/mbed_rtos_types.h"
 #include "rtos/internal/mbed_rtos1_types.h"
 #include "rtos/internal/mbed_rtos_storage.h"
 #include "platform/NonCopyable.h"
