+++ rtos/include/rtos/Queue.h
@@ -23,7 +23,7 @@
 #ifndef QUEUE_H
 #define QUEUE_H
 
-#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/mbed_rtos_types.h"
 #include "rtos/internal/mbed_rtos1_types.h"
 #include "rtos/internal/mbed_rtos_storage.h"
 #include "rtos/Kernel.h"
