+++ rtos/include/rtos/Semaphore.h
@@ -25,7 +25,7 @@
 
 #include <stdint.h>
 #include <chrono>
-#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/mbed_rtos_types.h"
 #include "rtos/internal/mbed_rtos1_types.h"
 #include "rtos/internal/mbed_rtos_storage.h"
 #include "rtos/Kernel.h"
