+++ rtos/include/rtos/Thread.h
@@ -24,7 +24,7 @@
 #define THREAD_H
 
 #include <stdint.h>
-#include "rtos/internal/mbed_rtos_types.h"
+#include "rtos/mbed_rtos_types.h"
 #include "rtos/internal/mbed_rtos1_types.h"
 #include "rtos/internal/mbed_rtos_storage.h"
 #include "platform/Callback.h"
